0.695
DONE
1. added two new logos
2. GUI resizable
3. define color CAPITALIZED / font define
4. changed rectification into a knob (original toggle buttons)
TODO
1. HQ / Link knob redesign
2. filter GUI redesign
3. visualiser
4. reduce CPU usage of oversampling
5. oversampling may sound different on diode mode.

+++ Source/Distortion.h
@@ -27,8 +27,8 @@ public:
         float output;
         // Mix, [0., 1.] ratio between a dry and wet signal
         float mix;
-        // Rectification mode 0 = bypass, 1 = half, 2 = full
-        int rectification;
+        // Rectification mode 0 = bypass, 0.5 = half, 1 = full
+        float rectification;
     } controls;
 
     Distortion();
